format money form float to int

+++ transaction.go
@@ -3,5 +3,5 @@ package main
 type Transaction struct {
 	Origin      string
 	Destination string
-	Value       float64
+	Value       int64
 }
